KAFKA-17047: Refactored group coordinator classes to modern package (KIP-932)
