+++ group-coordinator/src/main/java/org/apache/kafka/coordinator/group/modern/GroupSpecImpl.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.coordinator.group.consumer;
+package org.apache.kafka.coordinator.group.modern;
 
 import org.apache.kafka.common.Uuid;
 import org.apache.kafka.coordinator.group.api.assignor.GroupSpec;
@@ -28,7 +28,7 @@ import java.util.Map;
 import java.util.Objects;
 
 /**
- * The assignment specification for a consumer group.
+ * The assignment specification for a modern group.
  */
 public class GroupSpecImpl implements GroupSpec {
     /**
